+++ src/component/App.js
@@ -1,0 +1,14 @@
+import logo from '../logo.svg';
+import '../App.css';
+import Header from "./Header";
+
+function App() {
+  return (
+    <div className="App">
+      <Header />
+
+    </div>
+  );
+}
+
+export default App;
