+++ src/component/Header.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import logo from "../logo.svg";
+
+function Header() {
+    return(
+        <header className="App-header">
+            <img src={logo} className="App-logo" alt="logo" />
+            <p>
+                Edit <code>src/App.js</code> and save to reload.
+            </p>
+            <a
+                className="App-link"
+                href="https://reactjs.org"
+                target="_blank"
+                rel="noopener noreferrer"
+            >
+                Learn React js
+            </a>
+        </header>
+    )
+}
+
+export default Header;
